Scaffolding cleaning (#525)
* remove unnecessary cloning of component name

* move eval group to interpret_group file

* fix module structure

* reorganize files, remove interpreter structs, change many fns to use references. only parse once

* run validate before interpreting

* remove unneeded import

* cosmetic changes
