+++ interp/src/interpreter/mod.rs
@@ -1,0 +1,5 @@
+mod interpret_component;
+mod interpret_control;
+mod interpret_group;
+
+pub use interpret_component::interpret_component;
