+++ interp/src/lib.rs
@@ -1,7 +1,4 @@
 pub mod environment;
-pub mod interpret_component;
-pub mod interpret_control;
-pub mod interpret_group;
 pub mod interpreter;
 pub mod primitives;
 pub mod update;
